refactor(Board): rule 주입받도록 수정
