+++ src/test/kotlin/library/exactlyfive/ExactlyFiveTest.kt
@@ -1,7 +1,7 @@
 package library.exactlyfive
 
 import domain.CoordinateState
-import domain.rule.ExactlyFive
+import domain.library.cldhfleks2.ExactlyFive
 import org.assertj.core.api.Assertions.assertThat
 import org.junit.jupiter.params.ParameterizedTest
 import org.junit.jupiter.params.provider.CsvSource
