+++ src/test/kotlin/library/exceedfive/ExceedFiveTest.kt
@@ -1,7 +1,7 @@
 package library.exceedfive
 
 import domain.CoordinateState
-import domain.rule.ExceedFive
+import domain.library.cldhfleks2.ExceedFive
 import org.assertj.core.api.Assertions.assertThat
 import org.junit.jupiter.params.ParameterizedTest
 import org.junit.jupiter.params.provider.CsvSource
